requirements.txt
aiohttp
lxml
nonebot2
nonebot-adapter-onebot